Make Filter changes to ViolationGenerationEngine
